Fix: seeding more data and adding users
